fix: single ref keys changed to array of refs
courseId and AchievementObtainedId replaced with courses and achievements and made array of ObjectId schema type since users can have multiple courses, achievements and finishedCourses. Making array of objectId Schema type allow us to create one-to-many or many-to-many relations bc a schema can refer to multiple referance by storing more than one ObjectId.

Once we use populate method and pass the key to the method, we can get array of the object of the ref type which we want.